+++ spinetoolbox/widgets/settings_widget.py
@@ -71,6 +71,7 @@ class SettingsWidget(QWidget):
         self.ui.toolButton_bg_color.clicked.connect(self.show_color_dialog)
         self.ui.radioButton_bg_grid.clicked.connect(self.update_scene_bg)
         self.ui.radioButton_bg_solid.clicked.connect(self.update_scene_bg)
+        self.ui.checkBox_use_smooth_links.clicked.connect(self.update_links_geometry)
 
     @Slot(bool, name="browse_gams_path")
     def browse_gams_path(self, checked=False):
@@ -215,6 +216,14 @@ class SettingsWidget(QWidget):
             self._toolbox.ui.graphicsView.scene().set_bg_grid(False)
             self._toolbox.ui.graphicsView.scene().update()
 
+    @Slot(bool)
+    def update_links_geometry(self, checked=False):
+        from ..graphics_items import Link
+
+        for item in self._toolbox.ui.graphicsView.items():
+            if isinstance(item, Link):
+                item.do_update_geometry(checked)
+
     def read_settings(self):
         """Read saved settings from app QSettings instance and update UI to display them."""
         # checkBox check state 0: unchecked, 1: partially checked, 2: checked
@@ -225,6 +234,7 @@ class SettingsWidget(QWidget):
         datetime = int(self._qsettings.value("appSettings/dateTime", defaultValue="2"))
         delete_data = int(self._qsettings.value("appSettings/deleteData", defaultValue="0"))
         smooth_zoom = self._qsettings.value("appSettings/smoothZoom", defaultValue="false")
+        smooth_links = self._qsettings.value("appSettings/smoothLinks", defaultValue="false")
         bg_grid = self._qsettings.value("appSettings/bgGrid", defaultValue="false")
         bg_color = self._qsettings.value("appSettings/bgColor", defaultValue="false")
         gams_path = self._qsettings.value("appSettings/gamsPath", defaultValue="")
@@ -251,6 +261,8 @@ class SettingsWidget(QWidget):
             self.ui.checkBox_delete_data.setCheckState(Qt.Checked)
         if smooth_zoom == "true":
             self.ui.checkBox_use_smooth_zoom.setCheckState(Qt.Checked)
+        if smooth_links == "true":
+            self.ui.checkBox_use_smooth_links.setCheckState(Qt.Checked)
         if bg_grid == "true":
             self.ui.radioButton_bg_grid.setChecked(True)
         else:
@@ -314,6 +326,8 @@ class SettingsWidget(QWidget):
         self._qsettings.setValue("appSettings/deleteData", delete_data)
         smooth_zoom = "true" if int(self.ui.checkBox_use_smooth_zoom.checkState()) else "false"
         self._qsettings.setValue("appSettings/smoothZoom", smooth_zoom)
+        smooth_links = "true" if int(self.ui.checkBox_use_smooth_links.checkState()) else "false"
+        self._qsettings.setValue("appSettings/smoothLinks", smooth_links)
         bg_grid = "true" if self.ui.radioButton_bg_grid.isChecked() else "false"
         self._qsettings.setValue("appSettings/bgGrid", bg_grid)
         self._qsettings.setValue("appSettings/bgColor", self.bg_color)